+++ resources/js/focalpoint.js
@@ -1,5 +1,5 @@
-import FocalPointFieldtype from "./FocalPointFieldtype";
+import FocalPointFieldtype from './FocalPointFieldtype.vue';
 
 Statamic.booting(() => {
-    Statamic.$components.register('focalpoint-fieldtype', FocalPointFieldtype);
-})
+    Statamic.$components.register('focal_point-fieldtype', FocalPointFieldtype);
+});